requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
paddleocr==2.9.1
paddlepaddle==3.0.0b1
Pillow==11.0.0
opencv-python-headless==4.10.0.84
python-multipart==0.0.12
numpy==1.26.4
google-api-python-client==2.153.0
google-auth==2.36.0
pdfplumber==0.11.4
httpx>=0.27.0
